+++ components/Home/Epoch.tsx
@@ -11,6 +11,7 @@ import { Cell, Pie, PieChart, ResponsiveContainer, Tooltip } from 'recharts';
 import { BLOCK_SUBSCRIPTION } from '../Query/Block';
 import { GET_CHAIN } from '../Query/Chain';
 import { GET_EPOCH } from '../Query/Epoch';
+import ComponentLoader from '../Utils/ComponentLoader';
 import ErrorMessage from '../Utils/ErrorMessage';
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -129,7 +130,7 @@ const Epoch = (): JSX.Element => {
         averageBlockTime.data.chain.averageBlockTime
             ? averageBlockTime.data.chain.averageBlockTime
             : 0;
-    // if (loading) return <ComponentLoader />;
+    if (loading) return <ComponentLoader />;
     if (error)
         return (
             <ErrorMessage