+++ components/Transaction/LatestTransactions.tsx
@@ -17,6 +17,7 @@ import React, { useEffect } from 'react';
 
 import { GET_TX } from '../Query/Transaction';
 import Chips from '../Utils/Chips';
+import ComponentLoader from '../Utils/ComponentLoader';
 import ErrorMessage from '../Utils/ErrorMessage';
 import MiddleEllipsis from '../Utils/MiddleEllipsis';
 import NavLink from '../Utils/NavLink';
@@ -127,7 +128,7 @@ const LatestTransactions = ({ pagination }: LatestTxsProps): JSX.Element => {
         pollInterval: 5000
     });
 
-    // if (loading) return <ComponentLoader />;
+    if (loading) return <ComponentLoader />;
     if (error) return <ErrorMessage message={error.message} />;
 
     if (data && data.transactions)